requirements.txt
numpy>=1.26.0
piper-tts>=1.2.0
requests>=2.32.0
sounddevice>=0.5.0
